Upgrade s3o-middleware-utils to 2.0.0 (#53)
This removes express cookie utilities from s3o-middleware-utils, and exports a factory function from authenticate which takes a sync public key getter and returns authenticateToken and validate methods. This was to allow `s3o-lambda` to consume the same library without immediately polling for public key changes (https://github.com/Financial-Times/s3o-middleware-utils/pull/8).

This change adds the cookie utils to this library and adjusts the calls to the middleware utils to keep the existing API intact. I'd have liked to expose this middleware with a factory function as an entry point as well:

* Prevents immediate eager calls to the public key endpoint
* Makes it easier to tests APIs using this library without having to disable/stub s3o or do a conditional require

however I don't personally have the appetite to manage that upgrade process.

Note: s3o-middleware-utils now requires node.js >= 8. This is due to an ft-poller version update including native async/await calls. This should result in this being a major release. node.js 6 supports ends at the end of next month (April 2019) so I believe this is a reasonable breaking change at this point.

This would not require consumers to update immediately.