Cater for non-default ports (#56)
This respects non-default ports in the s3o redirect, which makes local
development with this middleware possible again. This fixes #54.

This is a little more code than I originally thought, because there's
not a very easy way to get a canonical port in Express. I'm using a mix
of the `x-forwarded-port` header and splitting the original hostname on
a colon.

I've written some automated tests, but I also checked that this works as
expected in a Heroku app - it does. Heroku specifies an
`x-forwarded-port` header of `443` which we then ignore.